did until ex4

+++ module_01/ex1/ft_garden_data.py
@@ -1,10 +1,10 @@
 class Plant:
-        def __init__(self, name: str, height: int, age: int) -> None:
-            self.name = name
-            self.height = height
-            self.age = age
-        def show(self) ->None:
-            print(f"{self.name}: {self.height}cm, {self.age} days old")
+    def __init__(self, name: str, height: float, age: int) -> None:
+        self.name = name
+        self.height = height
+        self.age = age
+    def show(self) ->None:
+        print(f"{self.name}: {self.height}cm, {self.age} days old")
 
 def ft_garden_data() -> None:
     print("=== Garden Plant Registry ===")
